installed geokit, added 1.6m uk postcodes into the seed db, modified postcode to be unique. One test is failing in rake for no apparent reason, passes on its own.

+++ app/controllers/user_sessions_controller.rb
@@ -11,7 +11,7 @@ class UserSessionsController < ApplicationController
   end
   
   def destroy
-    current_user_session.destroy
+    current_user_session.destroy if current_user_session
     flash[:notice] = "Looking forward to seeing you again!"
     redirect_back_or_to '/'
   end
